+++ __tests__/.eslintrc
@@ -1,0 +1,6 @@
+{
+  "extends": "aftership",
+  env: {
+    "jest": true
+  }
+}
